Use int instead of NSInteger which is not available on Tiger
git-svn-id: https://svn.macports.org/repository/macports/users/ryandesign/PCLauncher@73280 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ LoginWindowController.h
@@ -78,7 +78,7 @@
 
 - (void)showProgressPanel;
 - (void)setProgressPanelMessage:(NSString *)status;
-- (void)didEndSheet:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
+- (void)didEndSheet:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo;
 - (void)showAlertWithTitle:(NSString *)title message:(NSString *)message;
 
 - (NSString *)makeTempDirectory;
