+++ LoginWindowController.m
@@ -434,7 +434,7 @@ enum {
 	[progressLabel setStringValue:status];
 }
 
-- (void)didEndSheet:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo {
+- (void)didEndSheet:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo {
 	[sheet orderOut:self];
 }
